Updated the Robot class.
Added a tickable that updates the robot mode to the current mode as is dictated by the FMS/Driver Station. The Robot class now also imports constants from the PhysicalConstants class, which now holds all PWM/DIO/Solenoid channels used on the Robot. Build.java is removed, as the run target in build.xml is now defined as an empty target that depends on deploy, providing a much cleaner way of deploying.