Updated Roller behaviour.
The Roller class now behaves as requested by Carl Hansen. The rollers may always spin outwards, but each roller may only spin inwards if the limit switch on that side is not activated. The rollers will stop three seconds (or whatever the value of the constant ROLLER_TIME) after the last method call to out() or in().