Trimmed the fat from nbproject.
nbproject should now work on any system with wpilib in the user's home,
and now only contains the files it needs. More particularly,
project.properties has a classpath property that points to absolute
paths, rather than the path of a library which might not be added on a
particular system.